feat(devtools): migrate plugin and add run tool
