+++ packages/plugin-kit/package.json
@@ -21,6 +21,11 @@
       "import": "./dist/workspace.js",
       "require": "./dist/workspace.cjs"
     },
+    "./devtools": {
+      "types": "./dist/devtools.d.ts",
+      "import": "./dist/devtools.js",
+      "require": "./dist/devtools.cjs"
+    },
     "./src": {
       "types": "./src/index.ts",
       "import": "./src/index.ts"
@@ -32,6 +37,10 @@
     "./src/workspace": {
       "types": "./src/workspace/index.ts",
       "import": "./src/workspace/index.ts"
+    },
+    "./src/devtools": {
+      "types": "./src/devtools/index.ts",
+      "import": "./src/devtools/index.ts"
     }
   },
   "publishConfig": {
