+++ packages/plugin-kit/tsup.config.ts
@@ -5,6 +5,7 @@ export default defineConfig({
     index: 'src/index.ts',
     worktree: 'src/worktree/index.ts',
     workspace: 'src/workspace/index.ts',
+    devtools: 'src/devtools/index.ts',
   },
   format: ['esm', 'cjs'],
   dts: true,
